feat(register): se agregó el registro de usuario
