+++ marly-handmade-frontend/src/App.jsx
@@ -3,7 +3,7 @@ import Header from './components/Header';
 import reactLogo from './assets/react.svg'
 import viteLogo from '/vite.svg'
 import Footer from "./components/Footer";
-import Login from './pages/Login';
+import Register from './pages/Register';
 
 
 function App() {
@@ -13,7 +13,7 @@ function App() {
     <>
       <Header />
 
-      <Login /> 
+      <Register /> 
 
       <Footer />
     </>
